Update app.ts : 디렉토리 및 파일 분리
