+++ Lecture2_HackerNews_TS/index.html
@@ -20,6 +20,6 @@
     <noscript> You need to enable JavaScript to run this app </noscript>
 
     <div id="root"></div>
-    <script src="app.ts" type="module"></script>
+    <script src="src/app.ts" type="module"></script>
   </body>
 </html>
